# docs/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: 寝室电费学生端 API
  version: 0.7.0
  description: T0 冻结契约。金额为十进制字符串，未知为 null，日期为上海含首尾日期，时间戳带时区。202 仅表示持久受理。业务服务按 x-stage
    实现。
servers:
- url: /api/v1
paths:
...
  /monitor:
    get:
      operationId: get_monitor
      tags:
      - monitoring
      summary: get_monitor
      x-owner: monitoring
      x-stage: T3
      security:
      - SessionCookie: []
      parameters: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Monitor'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
    patch:
      operationId: update_monitor
      tags:
      - monitoring
      summary: update_monitor
      x-owner: monitoring
      x-stage: T3
      security:
      - SessionCookie: []
      parameters:
      - name: Origin
        in: header
        required: true
        schema:
          type: string
          format: uri
      - $ref: '#/components/parameters/CSRF'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Monitor'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitorPatch'
      x-version-required: expected_version；缺失 428，冲突 409
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-elect_session
      description: Secure; HttpOnly; SameSite=Lax; Path=/；不设置 Domain
  parameters:
    CSRF:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    Meta:
      additionalProperties: false
      properties:
        request_id:
          format: uuid
          title: Request Id
          type: string
        server_time:
          format: date-time
          title: Server Time
          type: string
      required:
      - request_id
      - server_time
      title: Meta
      type: object
    ErrorCode:
      enum:
      - INVALID_ARGUMENT
      - INVALID_INTERVAL
      - INVALID_DATE_RANGE
      - CAPTCHA_INVALID
      - CAPTCHA_EXPIRED
      - APP_SESSION_EXPIRED
      - SCHOOL_LOGIN_REJECTED
      - CSRF_REJECTED
      - ORIGIN_REJECTED
      - NOT_FOUND
      - VERSION_CONFLICT
      - IDEMPOTENCY_CONFLICT
      - OPERATION_IN_PROGRESS
      - SCHOOL_REAUTH_REQUIRED
      - REAUTH_ACCOUNT_MISMATCH
      - ROOM_CANDIDATE_EXPIRED
      - SCHOOL_BINDING_REJECTED
      - SNAPSHOT_EXPIRED
      - SNAPSHOT_MISMATCH
      - PRECONDITION_REQUIRED
      - RATE_LIMITED
      - SCHOOL_PROTOCOL_CHANGED
      - SCHOOL_INVALID_RESPONSE
      - SCHOOL_UNAVAILABLE
      - CIRCUIT_OPEN
      - DEPENDENCY_UNAVAILABLE
      - SCHOOL_TIMEOUT
      - PAYMENT_UNAVAILABLE
      - FEATURE_DISABLED
      title: ErrorCode
      type: string
    Error:
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          title: Message
          type: string
        retryable:
          title: Retryable
          type: boolean
        retry_after_seconds:
          anyOf:
          - minimum: 0
            type: integer
          - type: 'null'
          title: Retry After Seconds
        requires_reauth:
          title: Requires Reauth
          type: boolean
        field_errors:
          additionalProperties:
            type: string
          title: Field Errors
          type: object
        current_version:
          anyOf:
          - minimum: 1
            type: integer
          - type: 'null'
          default: null
          title: Current Version
        existing_operation_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          default: null
          title: Existing Operation Id
      required:
      - code
      - message
      - retryable
      - retry_after_seconds
      - requires_reauth
      - field_errors
      title: Error
      type: object
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/Error'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - error
      - meta
      title: ErrorEnvelope
      type: object
...
    MonitorConfig:
      additionalProperties: false
      properties:
        enabled:
          title: Enabled
          type: boolean
        interval_minutes:
          maximum: 1440
          minimum: 60
          title: Interval Minutes
          type: integer
        repeat_limit:
          maximum: 5
          minimum: 1
          title: Repeat Limit
          type: integer
        threshold:
          pattern: ^(0|[1-9]\d{0,11})\.\d{2}$
          title: Threshold
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
      required:
      - enabled
      - interval_minutes
      - repeat_limit
      - threshold
      - email
      title: MonitorConfig
      type: object
    MonitorPatch:
      additionalProperties: false
      properties:
        expected_version:
          minimum: 1
          title: Expected Version
          type: integer
        enabled:
          title: Enabled
          type: boolean
        interval_minutes:
          maximum: 1440
          minimum: 60
          title: Interval Minutes
          type: integer
        repeat_limit:
          maximum: 5
          minimum: 1
          title: Repeat Limit
          type: integer
        threshold:
          pattern: ^(0|[1-9]\d{0,11})\.\d{2}$
          title: Threshold
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: Email
      required:
      - expected_version
      title: MonitorPatch
      type: object
    NotificationSummary:
      additionalProperties: false
      properties:
        state:
          enum:
          - idle
          - pending
          - sending
          - sent
          - email_failed
          - delivery_unknown
          - cancelled
          title: State
          type: string
        last_sent_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Sent At
        last_error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error Code
        next_retry_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Next Retry At
        in_flight_count:
          minimum: 0
          title: In Flight Count
          type: integer
        delivery_unknown_count:
          minimum: 0
          title: Delivery Unknown Count
          type: integer
      required:
      - state
      - last_sent_at
      - last_error_code
      - next_retry_at
      - in_flight_count
      - delivery_unknown_count
      title: NotificationSummary
      type: object
    RunAttempt:
      additionalProperties: false
      properties:
        attempt_no:
          minimum: 1
          title: Attempt No
          type: integer
        started_at:
          format: date-time
          title: Started At
          type: string
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        outcome:
          enum:
          - running
          - succeeded
          - retryable
          - failed
          - cancelled
          - fenced
          title: Outcome
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
      required:
      - attempt_no
      - started_at
      - finished_at
      - outcome
      - error_code
      title: RunAttempt
      type: object
    Run:
      additionalProperties: false
      properties:
        id:
          format: uuid
          title: Id
          type: string
        version:
          minimum: 1
          title: Version
          type: integer
        binding_id:
          format: uuid
          title: Binding Id
          type: string
        state:
          enum:
          - pending
          - running
          - retry_wait
          - cancel_requested
          - succeeded
          - failed
          - cancelled
          title: State
          type: string
        scheduled_for:
          format: date-time
          title: Scheduled For
          type: string
        started_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        next_attempt_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Next Attempt At
        cancel_pending:
          title: Cancel Pending
          type: boolean
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        attempts:
          items:
            $ref: '#/components/schemas/RunAttempt'
          title: Attempts
          type: array
      required:
      - id
      - version
      - binding_id
      - state
      - scheduled_for
      - started_at
      - finished_at
      - next_attempt_at
      - cancel_pending
      - error_code
      - attempts
      title: Run
      type: object
...
    Monitor:
      additionalProperties: false
      properties:
        id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Id
        binding_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Binding Id
        config:
          $ref: '#/components/schemas/MonitorConfig'
        state:
          enum:
          - active
          - disabled
          - requires_reauth
          - blocked_room
          - retargeting
          title: State
          type: string
        health:
          enum:
          - healthy
          - degraded
          - unavailable
          title: Health
          type: string
        version:
          minimum: 1
          title: Version
          type: integer
        generation:
          minimum: 1
          title: Generation
          type: integer
        current_run:
          anyOf:
          - $ref: '#/components/schemas/Run'
          - type: 'null'
        last_run:
          anyOf:
          - $ref: '#/components/schemas/Run'
          - type: 'null'
        next_run_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Next Run At
        last_success_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Success At
        last_error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error Code
        cancel_pending:
          title: Cancel Pending
          type: boolean
        in_flight_count:
          minimum: 0
          title: In Flight Count
          type: integer
        notification:
          $ref: '#/components/schemas/NotificationSummary'
      required:
      - id
      - binding_id
      - config
      - state
      - health
      - version
      - generation
      - current_run
      - last_run
      - next_run_at
      - last_success_at
      - last_error_code
      - cancel_pending
      - in_flight_count
      - notification
      title: Monitor
      type: object
...